// src/enums/logical-symbols.ts
export enum LogicalSymbolHexCode {
  Implication = "\u21D2",
  Conjunction = "\u2227",
  Disjunction = "\u2228",
  Negation = "\xAC",
  Equivalence = "\u21D4",
}

export enum LogicalSymbolRawInput {
  Implication = "=>",
  Conjunction = "&",
  Disjunction = "|",
  Negation = "~",
  Equivalence = "<=>",
  OpenParenthesis = "(",
  CloseParenthesis = ")",
}

export enum CommonSymbolHexCode {
  Space = "\xA0",
}